requirements.txt
alchemyjsonschema==0.6.1
alembic==1.5.8
aniso8601==9.0.1
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
boto==2.49.0
boto3==1.17.38
botocore==1.20.38
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
datalake==0.33
datalake-common==0.26
decorator==4.4.2
dictknife==0.13.0
distlib==0.3.1
dnspython==2.1.0
docopt==0.6.2
email-validator==1.1.2
ERAlchemy==1.2.10
filelock==3.0.12
flake8==3.5.0
Flask==1.1.2
Flask-Cors==3.0.9
flask-request-id-header==0.1.1
Flask-RESTful==0.3.8
flask-restful-swagger-3==0.1
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gelf-formatter==0.2.0
greenlet==1.0.0
gunicorn==20.0.4
h11==0.12.0
identify==2.2.1
idna==2.10
infinity==1.5
inflect==5.3.0
intervals==0.9.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
magicalimport==0.9.1
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
memoized-property==1.0.3
mergedeep==1.3.4
mypy-extensions==0.4.3
nodeenv==1.5.0
openapi-schema-validator==0.1.5
packaging==20.9
pathspec==0.8.1
pre-commit==2.11.1
psycopg2==2.8.6
pyblake2==1.1.2
pycodestyle==2.3.1
pyflakes==1.6.0
pygraphviz==1.7
pykwalify==1.8.0
pyparsing==2.4.7
PyPika==0.48.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
ruamel.yaml==0.17.0
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
schemainspect==3.0.1616029793
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.4.3
sqlalchemy-filters==0.12.0
SQLAlchemy-Utils==0.36.8
sqlbag==0.1.1616397193
starlette==0.13.6
strict-rfc3339==0.7
swagger-ui==0.1.2
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
validators==0.18.2
virtualenv==20.4.3
Werkzeug==0.16.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5
WTForms-JSON==0.3.3
